Add files via upload
FLAMES is a popular game among budding teenagers and people in their early 20's which is played to find out the hypothetical relationships between 2 people. It works by crossing out the pair of equal alphabets from the names of the subjected people. The total number of unequal alphabets are counted and it is used as a counter to calculate the surviving letter from the word FLAMES. 
F : FRIENSHIP
L : LOVER
A : AFFECTION
M : MARRIAGE
E : ENEMITY
S : SISTER

!! IT'S A FUN GAME !!