Add files via upload
FLAMES is a popular game among the buddies and teenagers in their early 20's. It hypothetically describes the relationships between two based on their names. Crossing out all the similar letters in common in the names and the total count of the remaining letters  become the key to the game. The key number iterates over the word FLAMES and over repeated iteration it returns a single letter which comes with proper meaning behind. 
F : FRIENDSHIP
L : LOVER
A : AFFECTION
M : MARRIAGE
E : ENEMITY
S : SISTER


!! IT IS A FUN GAME !!